requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.13
cython==3.2.1
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
fastapi==0.122.0
filelock==3.20.3
flask==3.1.3
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
g2pkk==0.1.2
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.0
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markdown==3.10
markupsafe==3.0.3
matplotlib==3.10.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.4
num2words==0.5.14
numba==0.62.1
numpy==1.26.4
ollama==0.6.1
packaging==25.0
pandas==1.5.3
pillow==12.2.0
platformdirs==4.5.0
playsound3==3.3.0
pooch==1.8.2
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.1.3
pyaudio==0.2.14
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.33.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smart-open==7.5.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechrecognition==3.14.4
srsly==2.5.2
srt==3.5.3
starlette==0.50.0
sudachidict-core==20251022
sudachipy==0.6.10
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
trainer==0.0.36
transformers==5.0.0rc3
triton==3.4.0
tts==0.22.0
typeguard==4.4.4
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
unidecode==1.4.0
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
werkzeug==3.1.6
wheel==0.46.2
wrapt==2.0.1
yarl==1.22.0
langchain
langchain-mcp-adapters
langchain-ollama
langgraph
mcp
openmeteo-requests
piper-tts
python-dotenv==1.0.1
requests-cache
retry-requests
torch==2.8.0
torchaudio==2.8.0
urllib3==2.6.3
